moved CustomerProjector to PgcUowProjectorIT
